refactor(options): remove unimplemented logServer option
