--- docs/options/logServer.md
@@ -1,11 +1,0 @@
----
-"#type": "[[option]]"
-short: l
-type: "[[boolean]]"
-summary: log inner server traces
-batchForwarded: yes
-tags:
-  - legacy
-  - remote
-  - batch
----
