+++ docs/options/option.md
@@ -21,7 +21,6 @@ Options can be used in different places :
 | [[failFast]]             | f     | Stop the execution after the first failing page                                          | boolean         |          |                             |                           | 💻🔗🧪📡   |
 | [[failOpaFast]]          | fo    | Stop the OPA page execution after the first failing test                                 | boolean         |          |                             |                           | 💻🔗🧪📡   |
 | [[keepAlive]]            | k     | Keep the server alive                                                                    | boolean         |          |                             |                           | 💻🔗🧪📡   |
-| [[logServer]]            | l     | Log inner server traces                                                                  | boolean         |          |                             |                           | 💻🔗🧪📡   |
 | [[parallel]]             | p     | Number of parallel tests executions                                                      | integer         |          | 2                           |                           | 💻🔗🧪📡   |
 | [[browser]]              | b     | Browser instantiation command (relative to cwd or use $/ for provided ones)              | file            |          | '$/puppeteer.js'            |                           | 💻🔗🧪📡   |
 | [[alternateNpmPath]]     |       | Alternate NPM path to look for packages (priority: local, alternate, global)             | folder          |          |                             |                           | 💻🔗📡     |
